storage: add file upload/download methods
Summary:
This renames the existing upload_file and download_file to be upload_blob and download_blob.

It then adds upload_file and download_file methods which can be used to download directly to disk without having to pull it into a python in-memory representation. This will be helpful when interacting with torch lightning/classy vision which save/load models from files on disk.

Reviewed By: tierex

Differential Revision: D28456634

fbshipit-source-id: cb03273495dec0e58afb44110012df88a4a770f1