fix tests + torchx CLI (#6)
Summary:
This fixes the tests so they pass on OSS + make it so torchx entry point is correctly setup.

It also fixes a lot of imports to point to the torchx copy so it works in OSS since torchelastic is out of date there.

Pull Request resolved: https://github.com/pytorch/torchx/pull/6

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

  buck test //torchx/...

run OSS CI

Reviewed By: tierex

Differential Revision: D28614651

Pulled By: d4l3k

fbshipit-source-id: f8cd9577ac8ec2dbadf73d3539a4cbeb31b76bfa