kfp integration test (#21)
Summary:
adds a KFP integration test

This is a bit weird since KFP doesn't have a good way to do programmatic API access on AWS. We end up login in using the KFP cluster username/password and passing the cookie to the KFP client as endorsed by the docs.

Secrets are only provided for PRs that are run from the main repo as a branch and not forks so this will fail on forked PRs.

Pull Request resolved: https://github.com/pytorch/torchx/pull/21

Test Plan:
```
scripts/kfpint.py
```

http://5ab6bab9-istiosystem-istio-2af2-1926929629.us-west-2.elb.amazonaws.com/_/pipeline/#/runs/details/f7c56bef-909e-4660-b4c7-eeef486b224f

Reviewed By: kiukchung

Differential Revision: D28816114

Pulled By: d4l3k

fbshipit-source-id: b25f92cfc0af431d6b1b3a9e61331fe8f91c8c39