Introduce new Resource Definition API (#41)
Summary:
Pull Request resolved: https://github.com/pytorch/torchx/pull/41

The diff introduces a string representation of the resource definitions
Users can use these predefined resources in the following way:

    import torch_dist_role

    torch_dist_role(name="trainer", resource="TC", ..)

The diff also adds support for user-defined resources via entrypoints.
Users can define entrypoints in format:

    [torchx.base]
    resource_defs = custom_path:list_resources

The FB specific resources are defined in ``torchx.components.base.fb.resource_def`` . The resource.cpu is defined as a number of cores.

The diff also adds resource resolution for both mast and flow scheduler: adjusting memory to memory tax and scaling cpu(in flow case) to set it as number of hyperthreads.

Reviewed By: d4l3k

Differential Revision: D28953775

fbshipit-source-id: bd0b2c37b5a283f0bd88ca74cc35c9be761b453e