docs: added versions index (#49)
Summary:
This adds a link to versions.html as well as generates the versions.html using the `tree` command. It's not super pretty but it's easy to modify the CSS down the line.

I've also modified the script to push to the current origin so forks push to their own fork gh-pages.

Pull Request resolved: https://github.com/pytorch/torchx/pull/49

Test Plan:
```
./doc_push.sh
```
https://d4l3k.github.io/torchx/versions.html

Reviewed By: kiukchung

Differential Revision: D29046172

Pulled By: d4l3k

fbshipit-source-id: 0617db1bed437cc3b5794e67fa8c0933a08163c6