pipelines/kfp: add UI metadata support for Visualizations (#74)
Summary:
This adds support for KFP UI metadata to show metrics and outputs from the operator.

To avoid any container changes, we run a sidecar attached to the pod that uses a shared emptyDir volume mount on /tmp/ to write the metadata out as a KFP output. KFP outputs are always placed in /tmp/outputs/. This works reasonably well and is completely transparent to the running container (except for the /tmp mount which should be a no-op unless they're doing something very very strange).

If we add torchx.tracker support we could use it similarly to show input/outputs in the KFP UI.

More about the UI metadata: https://www.kubeflow.org/docs/components/pipelines/sdk/output-viewer/

Pull Request resolved: https://github.com/pytorch/torchx/pull/74

Test Plan:
http://5ab6bab9-istiosystem-istio-2af2-1926929629.us-west-2.elb.amazonaws.com/_/pipeline/#/runs/details/48c28c19-40a7-4914-aaa8-db2315210079

![2021-06-17-124548_1167x533_scrot](https://user-images.githubusercontent.com/909104/122463070-09a19700-cf6a-11eb-92bc-71a3ae7e50d9.png)

There's something weird with our cluster setup, the exact same tensorboard path runs locally but not in K8S. I think it might be a tensorboard 2.0.0 vs 2.4.1 issue or something to do with our test cluster setup. This is how it looks locally with the exact same path:
![2021-06-17-124534_745x666_scrot](https://user-images.githubusercontent.com/909104/122463081-0a3a2d80-cf6a-11eb-9dba-c569d40c1d12.png)

Reviewed By: tierex

Differential Revision: D29207442

Pulled By: d4l3k

fbshipit-source-id: 92961588f2f4fbc1d7ec31f6c32bd15e0c99bd91