adapters/kfp: support distributed training (#109)
Summary:
This adds a new `resource_to_app` KFP adapter that allows adapting an app to a kfp ResourceOp that launches the operator using the Volcano scheduler. This reuses the same code that creates the resources for the kubernetes scheduler and embeds the resource inside a KFP pipeline.

This isn't supported under KFP v2 since it interacts directly with kubernetes resources/volcano. This also requires volcano to be installed on the cluster to use which is why it's a new adapter instead of automatically being used.

This is still fairly experimental and once KFP has better distributed support we likely want to rely on that instead since this has some less than ideal UX. You need to use the CLI to access the individual worker logs and there isn't any support for UI metadata yet.

UI metadata I think can be added by providing an output annotation for argo as part of the resource but I haven't looked into it.

Pull Request resolved: https://github.com/pytorch/torchx/pull/109

Test Plan:
```
pyre
pytest
python dist_pipeline.py
```
http://5ab6bab9-istiosystem-istio-2af2-1926929629.us-west-2.elb.amazonaws.com/_/pipeline/#/runs/details/27707de9-bc67-42da-ab86-af2127ee54d1

![20210726_14h12m04s_grim](https://user-images.githubusercontent.com/909104/127059928-b4787429-e895-4b97-b53e-c6262e99c52b.png)

Reviewed By: kiukchung

Differential Revision: D29921246

Pulled By: d4l3k

fbshipit-source-id: ead6d0d4265cd22dd46fb89621a938c7281d7f5b