specs: relax definition of cpu to be a logical cpu core (#134)
Summary:
This updates the documentation to relax the definition of CPU core from physical core to logical core as the specific schedulers / cloud providers have varying definitions of it.

I.e. in Kubernetes:

> One cpu, in Kubernetes, is equivalent to 1 vCPU/Core for cloud providers and 1 hyperthread on bare-metal Intel processors.

https://kubernetes.io/docs/concepts/configuration/manage-resources-containers/#meaning-of-cpu

This makes it impossible for torchx to ensure that one cpu maps to a single physical core and thus we should document the behavior as such.

Pull Request resolved: https://github.com/pytorch/torchx/pull/134

Test Plan: no logic changes, CI

Reviewed By: aivanou

Differential Revision: D30286341

Pulled By: d4l3k

fbshipit-source-id: ba7111254744dfaf2de49834147b85d48b4b8b18