(bugfix)(torchx/runner) prevent cmd_run from stopping the app, add proper close methods to runner and scheduler instead
Summary:
A bug was introduced in https://github.com/pytorch/torchx/pull/141 that makes the `run` subcommand of the CLI always cancel (stop) the app in a finally block.

This diff introduces a proper `close()` method for `runner` and `scheduler` that is called to free/cleanup any stateful and allocated resources. Added a bunch of unittests and also a context manager so that `get_runner()` can now be used with `with`-statement that calls `close()` automagically.

Reviewed By: aivanou

Differential Revision: D30503364

fbshipit-source-id: 9f0c96de32bedef7c382bddc94e03bf2bbe6901d