components: copy component + app (#147)
Summary:
This adds a copy component and app as described in https://github.com/pytorch/torchx/issues/130.

If the file is the same filesystem it uses the .copy method to transfer the files, otherwise it read and writes it in 64 KB chunks.

Pull Request resolved: https://github.com/pytorch/torchx/pull/147

Test Plan:
```
$ pytest torchx/apps torchx/components
$ scripts/lint.sh
$ pyre
```

I haven't tested this as part of a E2E pipeline but will integrate it in the next diff to replace the data ingestion part of the advanced pipeline.

Reviewed By: kiukchung

Differential Revision: D30521666

Pulled By: d4l3k

fbshipit-source-id: 3a2816995d88d43fe2cd35b538cccfb1b6c4830d