Enable dist.ddp role to run on local scheduler (#137)
Summary:
Pull Request resolved: https://github.com/pytorch/torchx/pull/137

The changes the behavior of the `img_root` dir for local schedulers:
Instead of attaching `img_root` to the entrypoint and running it as :`img_root/entrypoint`, the diff adds the `img_root` to the PATH env variable. In this case the `entrypoint` will be available.

This change is necessary to allow running binaries that are not inside `img_root` dir , as well as binaries, that are inside the `img_root`

The diff also removes FbSpecific image type that is not used anymore

Reviewed By: kiukchung

Differential Revision: D30307869

fbshipit-source-id: 96ecb3a01ac7d98d4c88a5d3c270fd6db8914dd3