Add components integration test (#182)
Summary:
Pull Request resolved: https://github.com/pytorch/torchx/pull/182

The diff adds the components integration tests that current test all components on `local_docker` and `kubernetes` schedulers.

Add `--dryrun` to the example apps, to skip actual work
Add class that executes multiple AppDefs in a threadpool
The inte tests consist of builtin and example app definitions.
Since example app defs are outside of the torchx module namespace, we cannot put them in the `AppProvider`

Reviewed By: d4l3k

Differential Revision: D30980471

fbshipit-source-id: cadc6385641a3bc13683a764cfef59daa0ddf303