docs/configure: update entry_points docs + doctest (#221)
Summary:
The documentation for TorchX entry_points support was a bit outdated (as pointed out in T102501232). This updates it as well as switches it to use doctests to ensure that there's no regressions in the future.

Pull Request resolved: https://github.com/pytorch/torchx/pull/221

Test Plan:
```
make -C docs doctest
make -C docs livehtml
```
![Screenshot 2021-10-06 at 16-31-25 Configuring — PyTorch TorchX main documentation](https://user-images.githubusercontent.com/909104/136297399-9e072258-da7a-4658-9711-336201f21bd2.png)

Reviewed By: aivanou

Differential Revision: D31458545

Pulled By: d4l3k

fbshipit-source-id: 55a3e603f07ff4fba7c3ff7d4c1192b9d2f6b0ee