Introduce new `canary` command (#228)
Summary:
Pull Request resolved: https://github.com/pytorch/torchx/pull/228

The diff introduces fb-only `canary` that allows to execute component on mast scheduler without the need of rebuilding the package.

Fb `canary` is extension of `run` command that does patching. It tries to find a patching directory, copies it to the temporary location, compresses it into a tarball and puts it to the manifold.

It sets `PENV_PATCH` to the location of the manifold url.

The has the following assumptions:

* User fbpkg image should include `python` binary that is mapped to `//torchx/fb:python` .
* User execution par should be named `penv.par`. If users want to override this logic, they need to run torchx cmd with PENV_PARNAME='main.par' env variable
* `canary` currently does not have restrictions on the amount of files copied. It is up to user to make sure that the project is reasonable in size
* `canary` does NOT support projects where project modules are different from their location, e.g. targets have `base_module` defined
* `canary`  supports `--patch_dir`. When specified, the directory will be used, otherwise the cwd of command launch will be used.

The diff also adds ability to get commands from entrypoints.

Reviewed By: d4l3k

Differential Revision: D31296258

fbshipit-source-id: 0237481bd2ea11e0c3f513e29ce59a21556a7cb2