examples/lightning_classy_vision: switch model architecture to resnet so it actually converges (#245)
Summary:
This switches the TinyImageNetModel to use resnet with variable block sizes. To adjust model complexity the user can set `--layers` with basically any set of 4 ints in the ranges specified at:

https://github.com/pytorch/vision/blob/main/torchvision/models/resnet.py#L299

the example defaults to [1,1,1,1]
resnet18 is [2,2,2,2]
resnet34 is [3, 4, 6, 3]

This also adds learning rate to allow testing with that.

To get the model to converge you should run it with `--epochs 10 --num_samples 0` to do the full dataset with 10 passes. The whole dataset is quite large so running this on a GPU is ideal.

Pull Request resolved: https://github.com/pytorch/torchx/pull/245

Test Plan:
```
torchx run --scheduler local_cwd torchx/examples/apps/lightning_classy_vision/component.py:trainer --image ""
--data_path ~/Developer/tiny-imagenet-200/ --output /tmp/cv/ --log_path /tmp/logs --epochs 10 --num_samples 0
```

CI

Reviewed By: aivanou

Differential Revision: D31625999

Pulled By: d4l3k

fbshipit-source-id: 99bd2ad85b32a7109cdd53a3438e6a7defd5d022