Add the skeleton and boilerplate parts of Ray scheduler (#315)
Summary:
This is the initial PR of the TorchX Ray scheduler. It mostly contains the boilerplate parts of the scheduler, including the skeleton of the interface methods that should be implemented in the near future.

Note that this PR is a slight revision of https://github.com/pytorch/torchx/issues/219. I preferred to open a new PR since https://github.com/pytorch/torchx/issues/219 later grew to contain additional commits that were not meant to be part of the original PR.

Pull Request resolved: https://github.com/pytorch/torchx/pull/315

Reviewed By: aivanou, msaroufim

Differential Revision: D32082314

Pulled By: cbalioglu

fbshipit-source-id: 5afb33e8aa627cca80b2b291d1fad32244352c34