Replace calls to best point utilities in Ax tests (#367)
Summary:
Pull Request resolved: https://github.com/pytorch/torchx/pull/367

A problem arose because of versioning mismatch and backwards compatibility issues with the best_point.py utility methods and TorchX (in short, TorchX internally depends on the latest Ax which added a required parameter, but externally TorchX uses Ax stable release where the parameter is not present).

While this larger issue is being sorted out, I've removed the calls to the problematic method.

Reviewed By: lena-kashtelyan

Differential Revision: D33529812

fbshipit-source-id: 47f9af08bbc3e71d9db058002fb0cd77b187ea71