torchx/runner: log events to torch.monitor
Summary:
This logs the `torchx.runner.events.Events` to `torch.monitor` as well as the existing event handlers. Once monitor is stable the existing ones will be removed entirely in favor of the new interface.

`torch.monitor` is only available with pytorch 1.11 (or main) so it's a no-op if it's not available.

Differential Revision: D33928333

fbshipit-source-id: 632a7e096f46e5c1946932d4a765be179e1e53a8