torchx/runner: log events to torch.monitor (#379)
Summary:
This logs the `torchx.runner.events.Events` to `torch.monitor` as well as the existing event handlers. Once monitor is stable the existing ones will be removed entirely in favor of the new interface.

`torch.monitor` is only available with pytorch 1.11 (or main) so it's a no-op if it's not available. We import typestubs from 1.11 so pyre passes in OSS until we update to 1.11.

Pull Request resolved: https://github.com/pytorch/torchx/pull/379

Test Plan:
```
buck test //torchx/runner/...
```

Reviewed By: aivanou

Differential Revision: D33928333

Pulled By: d4l3k

fbshipit-source-id: dcb50fdf316db2f5243aa2564b022b1cc0770d86