(torchx/cli) allow --workspace to be configured through .torchxconfig
Summary: For `torchx run`, allows `--workspace` to be configured through `.torchxconfig`. Adds a custom `argparse.Action` that can be used to read and default other CLI options through `.torchxconfig`. Only hooked it up to `workspace` for now since there is some cleanup to be done to hook it up for `component`, `component args` (we get this from a different section not `[cli:run]`))

Differential Revision: D34317215

fbshipit-source-id: df09086bdb6778618f43f5cc534630dcb6193b1d