(torchx/cli) allow --workspace to be configured through .torchxconfig (#397)
Summary:
Pull Request resolved: https://github.com/pytorch/torchx/pull/397

For `torchx run`, allows `--workspace` to be configured through `.torchxconfig`. Adds a custom `argparse.Action` that can be used to read and default other CLI options through `.torchxconfig`. Only hooked it up to `workspace` for now since there is some cleanup to be done to hook it up for `component`, `component args` (we get this from a different section not `[cli:run]`))

Reviewed By: divchenko

Differential Revision: D34317215

fbshipit-source-id: 878f9d75002ab352f4d08d15802cfa405995f544