docker_workspace: support dockerignore to filter files from the workspace (#401)
Summary:
This adds a filter when building the docker workspace context to filter files specified in the `.dockerignore` file. This reduces the size of the patch when operating in a larger repo.

Reference:
https://docs.docker.com/engine/reference/builder/#dockerignore-file

Next steps from this would be to extend this logic to be workspace agnostic so we could have a `.torchxignore` file but not sure if that makes sense or is strictly necessary yet.

Pull Request resolved: https://github.com/pytorch/torchx/pull/401

Test Plan:
$  torchx run --scheduler aws_batch --wait --log dist.ddp --script foo.py  -j 1x1
$ pytest torchx/workspace/

Reviewed By: kiukchung

Differential Revision: D34486363

Pulled By: d4l3k

fbshipit-source-id: b71ac0bb2a798fdfd95f13703a700ab9850d127a