slurm_scheduler, dir_workspace: add isolated workspaces for Slurm (#416)
Summary:
This adds a new `DirWorkspace` which will copy the current workspace to the directory for code isolation purposes and integrates it in with the `slurm` scheduler via the `job_dir` runopt. The job dir must not exist and will be created. The CWD will be located in that job_dir.

* `.torchxignore` is used for excluding files from the workspace
* `.torchxslurmjobdirs` is used to track where job directories and thus logs are located

Pull Request resolved: https://github.com/pytorch/torchx/pull/416

Test Plan: Slurm integ tests + unit tests

Reviewed By: kiukchung

Differential Revision: D34801126

Pulled By: d4l3k

fbshipit-source-id: 92d98216d4f794604824b0bd1aceb6d0da4bae18