torchx/specs,schedulers: add new DeviceMount for mounting host devices (#442)
Summary:
This adds in a new mount type `DeviceMount` that allows mounting a host device into a container in the supported schedulers. Docker provides this as a special "devices" field but that generalizes to just mounting a host device into a container as you would with `mount --bind` so we call it a mount. It also avoids adding any new concepts to the CLI.

* Docker: sets the devices parameter https://docker-py.readthedocs.io/en/stable/containers.html#docker.models.containers.ContainerCollection.run
* AWS Batch: sets the devices parameter https://docs.aws.amazon.com/batch/latest/APIReference/API_LinuxParameters.html
* Kubernetes: does a bind mount w/ privileged mode since devices isn't supported https://stackoverflow.com/questions/59290752/how-to-use-device-dev-video0-with-kubernetes

Closes https://github.com/pytorch/torchx/issues/438

Pull Request resolved: https://github.com/pytorch/torchx/pull/442

Test Plan:
Added unit tests for docker, batch, k8s
CI
```
$ torchx run --scheduler aws_batch -c queue=torchx-p3dn --wait --log dist.ddp --memMB 16000 --env NCCL_DEBUG=INFO,LOGLEVEL=INFO,NCCL_SOCKET_IFNAME=eth0 -j 2x1 --image 763104351884.dkr.ecr.us-west-2.amazonaws.com/pytorch-training:1.10.2-gpu-py38-cu113-ubuntu20.04-e3 --mount type=device,src=/dev/infiniband/uverbs0 --gpu 8 --script dist_app.py
```

Reviewed By: kiukchung

Differential Revision: D35272170

Pulled By: d4l3k

fbshipit-source-id: 7793337cd83d5f16988c5b849fb1970f50f74525