(torchx/fb-image-provider)(bugfix) penv_parname should only be inferred if role.entrypoint is penv python, should equal role.entrypoint for everything else (#460)
Summary:
Pull Request resolved: https://github.com/pytorch/torchx/pull/460

Fixes this repro diff D35531168.

TL;DR when using `local_penv` with DPP (data preproc) roles, the dpp_master and dpp_worker roles also go through the `PenvImageProvider` to fetch the fbpkg. The issue is that `PenvImageProvider` was written prior to workspaces (jetter integration) and implicitly assumes that `role.entrypoint="python"` (penv python). So when a non-python entrypoint was specified there were unexpected behaviors.

Specificially running:

```
$ cd ~/fbcode/torchx/examples/apps
$ torchx run -- -j 1x1 -h t1 --num_dpp_workers=2 --num_dpp_masters=1
```

Results in the PenvImageProvider NOT downloading the fbpkgs for dpp_master and dpp_worker roles when there exists a `buck-out/opt/torchx/examples/apps/penv.par`, which happens if jetter decides to rebuild the fbpkg

Reviewed By: dracifer, aivanou

Differential Revision: D35636351

fbshipit-source-id: 80325bbd9bafe894063da80543ebae9cc9daab94