(torchx/local_scheduler) make INFO logging about log directory cleanup/preservation clearer + set torchelastic log_dir to the torchx base log dir to keep everything in one place
Summary:
Makes the INFO logging about the torchx log directory location clearer and adds suggestion on how to make the log directory persistent by passing `log_dir` scheduler cfg.

Also sets `PET_LOG_DIR` env var in local scheduler to make the torchelastic logs also go to the same directory as the torchx base log dir. This makes it in such a way that both are preserved or both are deleted to keep things in one place and consistent.

Reviewed By: d4l3k

Differential Revision: D36107583

fbshipit-source-id: a9fc421a2968aac645e85e55f93a3c628a0ef387