torchx: add metadata_urls to allow printing user links from the CLI (#532)
Summary:
Pull Request resolved: https://github.com/pytorch/torchx/pull/532

This adds a new `metadata_urls` to AppStatus and `metadata` to DescribeAppResponse. Metadata URLs are automatically returned from metadata values if they start with `http://` or `https://`. This isn't quite as clean as a dedicated URLs field that we discussed in https://github.com/pytorch/torchx/issues/509 but can work when calling status and not just during launch time.

This also enables a way for schedulers to return extra UI links and services that are scheduler specific but not available at launch time.

Differential Revision: D37353646

fbshipit-source-id: dfe629bf75b489a9c842d1bb1d45d0edaf9ce7f0