Get logs from running job with aws_batch scheduler (#542)
Summary:
Running jobs return information about the logStream in a different
location than completed jobs. In a running job, the `attempts` list is
empty. This patch checks the status of a job to pull the logStreamName
of a running batch job in order to return logs for it.

Pull Request resolved: https://github.com/pytorch/torchx/pull/542

Test Plan:
- Unit test
- Ran `torchx log` on a running aws batch job, and verified that logs are now returned.

Reviewed By: priyaramani, kurman

Differential Revision: D37805006

Pulled By: d4l3k

fbshipit-source-id: 409b91a088aaf98fae3cc2def35c7b2aee7b09d1