local_scheduler: better formating when using pdb (#563)
Summary:
This improves the log behavior when running interactive applications such as pdb for debugging purposes. It changes it so that the local scheduler loggers will read whatever is in the buffer instead of requiring a new line character. This means that when using pdb the REPL `(Pdb)` will be printed correctly.

There's a few additional tweaks to generally be more robust to handling new lines in depth.

Pull Request resolved: https://github.com/pytorch/torchx/pull/563

Test Plan:
Added unit tests + existing tests
```
$ torchx run -s local_cwd --log --wait utils.python -c "import pdb; pdb.set_trace()"
```
![2022-07-19-140514_622x202_scrot](https://user-images.githubusercontent.com/909104/179848282-bfa44618-1cc2-4527-a748-cce4a51ecd95.png)

Reviewed By: kurman

Differential Revision: D37972170

Pulled By: d4l3k

fbshipit-source-id: 2534926acc3f128f971983fce35d3994467c3b25