fix warning B908(2)
Fix the warning on the code:
B908: Contexts with exceptions assertions like with self.assertRaises or with pytest.raises should not have multiple top-level statements. Each statement should be in its own context. That way, the test ensures that the exception is raised only in the exact statement where you expect it.
This warning is raised by flake8-bugbear: https://github.com/PyCQA/flake8-bugbear