Make torchx scheduler opts support enums (#870)
Summary:

Added Support for Enumerations in scheduler options.

This is a bit more generic as it takes in a creator function which converts a CfgVal to another type. There are some limitations on how general it can be based on how the typing is setup. This also makes it tricky to make a convenience function to handle only Enums.

Differential Revision: D55551233