fix torchx runner schedule mock in torchx runner api test
Summary:
`test_run_from_module` mocks `Runner.schedule` but does not give it a return value.

so, when these tests run and call functions that read the return value of `schedule` an exception occurs because the return value is a MagicMock object wrapping schedule(), so any function expecting to read a string/AppHandle returned from `schedule` fails.

this diff adds a sample app handle return value.

this issue was discovered when adding the `log_event` context to `run_component`, which requires the return value of `schedule` to be read (D56073276)

Differential Revision: D56191855