Make TTFB torchx event aware of distributed_ai_stack
Summary: Add an optional field `distributed_ai_stack` into TorchX Event. This is a common and valuable field which is also existing in MAST job metadata, as well as needed for post data processing of AI Instrumentation Framework in order to include pre MAST job submission as part of TTFB measurement.

Differential Revision: D61578121