Add appdef metadata to torchx event (#947)
Summary:
Pull Request resolved: https://github.com/pytorch/torchx/pull/947

Add appdef metadata to torchx event class. Scuba logs does not get modified yet.
The motivation of adding this is that so far we need to log distributed_ai_stack which is part of the App Metadata into AI Instrumentation framework which can be used for post data process for training jobs of various training stacks, such as MVAI, Pyper, Conda_On_MAST etc.

Reviewed By: andywag

Differential Revision: D61632621