Remove experimental hot spare policy (#948)
Summary:
Pull Request resolved: https://github.com/pytorch/torchx/pull/948

This diff removes the experimental hot spare restart policy and uses role restart instead with quorum hosts set to the min nodes requirement. This isnt the best as this gives us no good way to differentiate between elasticity and quorum based restarts in the future but we can address this by supporting quorum restarts differently in the future.

Differential Revision: D61746435