sync torchx .pyre_configuration.internal with external config and upg… (#986)
Summary:
There are some issues with Pyre versions in github CI, and at the moment we aren't getting type errors.

Making a PR to
(a) bump to Ubuntu 24.04, because the root cause is a glibc issue
(b) upgrade to the  2024-11-25 release
(c) suppress errors

I'm making a fresh PR because the original one doesn't seem to trigger github CI (maybe because it was exported from a diff)


Test Plan: Ran github CI, the problem is fixed.

Reviewed By: jesszzzz

Differential Revision: D66994745

Pulled By: stroxler