warn on unknown runopt passed via command-line
Summary:
When you pass an invalid scheduler arg via the .torchxconfig file, we get a warning about it.

This change adds a similar warning when the scheduler arg is passed via a command-line argument.

(background: I spent a few hours yesterday debugging a torchx command for a job until spotting a typo in one of the arguments)

Differential Revision: D70562266