Fix broken docs-build by pinning protobuf-3.20.x (#1033)
Summary:
Pull Request resolved: https://github.com/pytorch/torchx/pull/1033

Pin protobuf-3.20.x (required by kfp-1.8.22) in dev-requirements and first install docs/requirements.txt then downgrade protobuf to 3.20.x.

What we oughta do is to upgrade to `kfp-2.x` but that's a much more intrusive change as we are bumping major versions. So pin to `apache-airflow==2.6.3` so that protobuf doesn't conflict with kfp.

Reviewed By: tonykao8080

Differential Revision: D72016216