expose runner itself with context manager (#1069)
Summary:
Right now, when we `enter` context after `get_fb_runner`, we are using the existing definition in `Runner` which causes type erasure to the super type `Runner`.
While this is fine for most cases, we lose the ability to access the apis in sub classes.

Here, we expose the self type instead to prevent erasing types

Reviewed By: kiukchung

Differential Revision: D72677312

Pulled By: bobyangyf